강원대 BE_조준환 4주차 과제 (1, 2, 3단계) (#302)
* feat(CategoryEntity.java): 카테고리 엔티티를 생성함

* feat(Category): 카테고리 DTO 추가 및 맵퍼 클래스 추가

* feat(Category): 카테고리 DTO 추가 및 맵퍼 클래스 추가 완료
요구 사항이 이름 뿐만 아니라 추가적으로 있던 점을 반영함.

* feat(Category): 카테고리 기능 추가 완료

* feat(ProductOptionDTO추가): ProductOptionDTO 추가 완료

* feat(ProductOption 엔티티 추가): ProductOption 엔티티 추가 완료

* feat(ProductOption 레포지트리 추가): 카테고리 레포지트리 추가 완료

* feat(ProductOption): 엔티티 클래스에 엔티티임을 명시함

* feat(ProductOption): 연관관계 추가

* feat(ProductOptionMapper): 옵션에 대한 맵퍼 클래스 추가

* refactor

* feat(WishList): product엔티티에 wishlist를 추가하도록 변경

* fix

* 동기화용 커밋:
1. 리스너 클래스 추가 예정
2. 옵션 서비스에 옵션 추가/업데이트 기능 구성중

* feat(ProductOption): 리스너 클래스 추가 및 업데이트 기능 완성

* fix(ProductOption): 상품 유무 확인 방식 변경

* test(ProductOption): 테스트 코드 추가

+++ src/main/java/gift/service/ProductService.java
@@ -1,25 +1,34 @@
 package gift.service;
 
 import gift.DTO.ProductDTO;
-import gift.aspect.CheckProductExists;
+import gift.DTO.ProductOptionDTO;
+import gift.entity.ProductEntity;
+import gift.exception.ProductNotFoundException;
 import gift.mapper.ProductMapper;
+import gift.mapper.ProductOptionMapper;
+import gift.repository.ProductOptionRepository;
 import gift.repository.ProductRepository;
-import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
 public class ProductService {
 
+    @Autowired
     ProductRepository productRepository;
 
+    @Autowired
     ProductMapper productMapper;
 
-    public ProductService(ProductRepository productRepository, ProductMapper productMapper) {
-        this.productRepository = productRepository;
-        this.productMapper = productMapper;
-    }
+    @Autowired
+    private ProductOptionMapper productOptionMapper;
+
+    @Autowired
+    private ProductOptionRepository productOptionRepository;
 
     /**
      * 새 상품을 생성하고 맵에 저장함
@@ -38,11 +47,17 @@ public class ProductService {
      *
      * @param id 조회할 상품의 ID
      */
-    @CheckProductExists
     @Transactional(readOnly = true)
     public ProductDTO getProduct(Long id) {
-        var productEntity = productRepository.findById(id).get();
+        var productEntity = productRepository.findById(id)
+            .orElseThrow(()-> new ProductNotFoundException("상품이 존재하지 않습니다."));
         return productMapper.toProductDTO(productEntity);
+    }
+
+    @Transactional(readOnly = true)
+    public ProductEntity getProductEntity(Long id) {
+        return productRepository.findById(id)
+            .orElseThrow(()-> new ProductNotFoundException("상품이 존재하지 않습니다."));
     }
 
     /**
@@ -59,9 +74,11 @@ public class ProductService {
      *
      * @param id 삭제할 상품의 ID
      */
-    @CheckProductExists
     @Transactional
     public void deleteProduct(Long id) {
+        if (!isProdutExit(id)) {
+            throw new ProductNotFoundException("상품이 존재하지 않습니다.");
+        }
         productRepository.deleteById(id);
     }
 
@@ -70,17 +87,25 @@ public class ProductService {
      *
      * @param productDTO 갱신할 상품 객체
      */
-    @CheckProductExists
+
     @Transactional
     public ProductDTO updateProduct(Long id, ProductDTO productDTO) {
+        if (!isProdutExit(id)) {
+            throw new ProductNotFoundException("상품이 존재하지 않습니다.");
+        }
         var productEntity = productMapper.toProductEntity(productDTO);
         productEntity.setId(id);
         productRepository.save(productEntity);
         return productMapper.toProductDTO(productEntity);
     }
 
+    public void updateProductEntity(Long id, ProductEntity productEntity) {
+        productEntity.setId(id);
+        productRepository.save(productEntity);
+    }
+
     /**
-     * 사용자 ID를 통해 사용자의 상품 목록을 가져옵니다.
+     * 사용자의 상품 목록을 가져옵니다.
      *
      * @param pageable 페이지 정보
      * @return ProductDTO 목록
@@ -88,6 +113,73 @@ public class ProductService {
     @Transactional(readOnly = true)
     public List<ProductDTO> getAllProducts(Pageable pageable) {
         var productEntities = productRepository.findAll(pageable).toList();
-        return productEntities.stream().map(productMapper::toProductDTO).toList();
+        return productEntities.stream()
+                .map(productMapper::toProductDTO).toList();
+    }
+
+    /**
+     * 상품 옵션을 반환함
+     *
+     * @param id 옵션을 조회할 상품의 ID
+     */
+    @Transactional(readOnly = true)
+    public List<ProductOptionDTO> getProductOptions(Long id){
+        var productEntity = productOptionRepository.findById(id).get().getProductEntity();
+        return productEntity.getProductOptions().stream()
+                .map(productOptionMapper::toProductOptionDTO).toList();
+    }
+
+    /**
+     * 상품 옵션을 추가함
+     *
+     * @param productOptionDTO 추가할 상품 옵션 객체
+     * @param productId        상품 옵션을 추가할 상품의 ID
+     */
+    @Transactional
+    public ProductOptionDTO addProductOption(Long productId, ProductOptionDTO productOptionDTO){
+        var productEntity = productRepository.findById(productId)
+            .orElseThrow(() -> new IllegalArgumentException("상품이 존재하지 않습니다."));
+
+        if (productEntity.getProductOptions().stream()
+            .anyMatch(option -> option.getName().equals(productOptionDTO.name()))) {
+            throw new IllegalArgumentException("이미 존재하는 옵션입니다.");
+        }
+
+        var productOptionEntity = productOptionMapper.toProductOptionEntity(productOptionDTO, productEntity, false);
+        productOptionRepository.save(productOptionEntity);
+
+        return productOptionMapper.toProductOptionDTO(productOptionEntity);
+    }
+
+    /**
+     * 상품 옵션을 삭제함
+     *
+     * @param id 삭제할 상품 옵션의 ID
+     */
+    @Transactional
+    public void deleteProductOption(Long id){
+        if (!isProdutExit(id)) {
+            throw new ProductNotFoundException("옵션이 존재하지 않습니다");
+        }
+        productOptionRepository.deleteById(id);
+    }
+
+    /**
+     * 상품 옵션을 갱신함
+     *
+     * @param id               갱신할 상품 옵션의 ID
+     * @param productOptionDTO 갱신할 상품 옵션 객체
+     */
+    @Transactional
+    public void updateProductOption(Long id, ProductOptionDTO productOptionDTO) {
+        var productOptionEntity = productOptionRepository.findById(id)
+            .orElseThrow(() -> new ProductNotFoundException("옵션이 존재하지 않습니다"));
+        productOptionEntity.setName(productOptionDTO.name());
+        productOptionEntity.setQuantity(productOptionDTO.quantity());
+        productOptionRepository.save(productOptionEntity);
+    }
+
+    public boolean isProdutExit(Long id) {
+        return productRepository.existsById(id);
     }
 }
